NO-TICKET: restructure execution-engine/storage
