+++ execution-engine/shared/Cargo.toml
@@ -12,5 +12,6 @@ hostname = "0.1.5"
 lazy_static = "1.3.0"
 libc = "0.2.54"
 log = "0.4.6"
+num = { version = "0.2.0", default-features = false }
 serde = { version = "1.0.90", features = ["derive"] }
 serde_json = "1.0.39"
