+++ execution-engine/shared/src/lib.rs
@@ -1,4 +1,5 @@
-#![feature(never_type)]
+#![feature(result_map_or_else, never_type)]
+
 extern crate blake2;
 extern crate chrono;
 extern crate common;
@@ -6,10 +7,13 @@ extern crate common;
 extern crate lazy_static;
 extern crate libc;
 extern crate log;
+extern crate num;
 
+pub mod init;
 pub mod logging;
 pub mod newtypes;
 pub mod os;
 pub mod semver;
 pub mod socket;
 pub mod test_utils;
+pub mod transform;
