+++ execution-engine/storage/Cargo.toml
@@ -6,7 +6,6 @@ authors = ["Michael Birch <birchmd@casperlabs.io>"]
 [dependencies]
 failure = "0.1.5"
 lmdb = "0.8.0"
-num = { version = "0.2.0", default-features = false }
 wasmi = "0.4.2"
 common = { path = "../common", features = ["std", "gens"], package = "casperlabs-contract-ffi" }
 parking_lot = "0.7.1"
