+++ execution-engine/storage/src/lib.rs
@@ -1,28 +1,30 @@
 #![feature(never_type, result_map_or_else)]
 
+// core dependencies
+extern crate core;
+
+// third-party dependencies
 #[macro_use]
 extern crate failure;
-
-extern crate common;
 extern crate lmdb;
-extern crate num;
 extern crate parking_lot;
-extern crate shared;
 extern crate wasmi;
 
+// local dependencies
+extern crate common;
+extern crate shared;
+
+// third-party dev-dependencies
 #[cfg(test)]
 #[macro_use]
 extern crate lazy_static;
-
 #[cfg(test)]
 extern crate proptest;
-
-extern crate core;
 #[cfg(test)]
 extern crate tempfile;
 
+// modules
 pub mod error;
 pub mod global_state;
-pub mod history;
-pub mod op;
-pub mod transform;
+pub mod trie;
+pub mod trie_store;
